+++ src/OpenLoco/Vehicles/Vehicle.h
@@ -3,6 +3,7 @@
 #include "../Audio/Audio.h"
 #include "../Company.h"
 #include "../Entities/Entity.h"
+#include "../Objects/ObjectManager.h"
 #include "../Objects/VehicleObject.h"
 #include "../Speed.hpp"
 #include "../Types.hpp"
@@ -12,6 +13,8 @@
 
 namespace OpenLoco::Vehicles
 {
+    using CargoTotalArray = std::array<uint32_t, ObjectManager::getMaxObjects(ObjectType::cargo)>;
+
     constexpr auto kMaxVehicleLength = 176; // TODO: Units?
 
     void create(OpenLoco::Interop::registers& regs);
@@ -312,6 +315,8 @@ namespace OpenLoco::Vehicles
         OrderRingView getCurrentOrders() const;
         bool isPlaced() const { return tile_x != -1 && !(var_38 & Flags38::isGhost); }
         char* generateCargoTotalString(char* buffer);
+        char* generateCargoCapacityString(char* buffer);
+        char* cargoLUTToString(CargoTotalArray& cargoTotals, char* buffer);
         bool canBeModified() const;
         void liftUpVehicle();
         void sub_4B7CC3();